fix service managegeoinfor

+++ src/main/java/capstone/lip/landinformationportal/service/SegmentOfStreetService.java
@@ -1,12 +1,18 @@
 package capstone.lip.landinformationportal.service;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import capstone.lip.landinformationportal.entity.FormedCoordinate;
+import capstone.lip.landinformationportal.entity.RealEstateAdjacentSegment;
 import capstone.lip.landinformationportal.entity.SegmentOfStreet;
 import capstone.lip.landinformationportal.repository.SegmentOfStreetRepository;
+import capstone.lip.landinformationportal.service.Interface.IFormedCoordinate;
+import capstone.lip.landinformationportal.service.Interface.IRealEstateAdjacentSegmentService;
 import capstone.lip.landinformationportal.service.Interface.ISegmentOfStreetService;
 
 @Service
@@ -27,7 +33,13 @@ public class SegmentOfStreetService implements ISegmentOfStreetService {
 	@Override
 	public boolean delete(List<SegmentOfStreet> listSegment) {
 		try {
-			segmentOfStreetRepository.deleteInBatch(listSegment);
+			if (listSegment == null) throw new Exception("List segment is null");
+			if (listSegment.isEmpty()) throw new Exception("List segment is empty");
+			List<FormedCoordinate> listCoordinate = listSegment.stream().map(x->x.getListFormedCoordinate()).flatMap(List::stream).collect(Collectors.toList());
+			coordinateService.delete(listCoordinate);
+			List<RealEstateAdjacentSegment> listAdj = listSegment.stream().map(x->x.getListRealEstateAdjacentSegment()).flatMap(List::stream).collect(Collectors.toList());
+			adjService.delete(listAdj);
+			segmentOfStreetRepository.deleteAll(listSegment);
 			return true;
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -35,15 +47,40 @@ public class SegmentOfStreetService implements ISegmentOfStreetService {
 		}
 	}
 
+	@Autowired
+	private IFormedCoordinate coordinateService;
+	
+	@Autowired
+	private IRealEstateAdjacentSegmentService adjService;
+	
 	@Override
 	public boolean delete(SegmentOfStreet segmentOfStreet) {
 		
 		try {
+			if (segmentOfStreet == null) throw new Exception("Segment is null");
+			List<FormedCoordinate> listCoordinate = segmentOfStreet.getListFormedCoordinate();
+			coordinateService.delete(listCoordinate);
+			List<RealEstateAdjacentSegment> listAdj = segmentOfStreet.getListRealEstateAdjacentSegment();
+			adjService.delete(listAdj);
+			
 			segmentOfStreetRepository.delete(segmentOfStreet);
 			return true;
 		} catch (Exception e) {
 			e.printStackTrace();
 			return false;
+		}
+	}
+	@Override
+	public SegmentOfStreet findById(Long id) {
+		try {
+			Optional<SegmentOfStreet> segment = segmentOfStreetRepository.findById(id);
+			if (segment.isPresent()) {
+				return segment.get();
+			}
+			return null;
+		} catch (Exception e) {
+			e.printStackTrace();
+			return null;
 		}
 	}
